feat: separate out signal and node tests
